% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfrail_aux.R
\name{dist_to_pars}
\alias{dist_to_pars}
\title{dist_to_pars}
\usage{
dist_to_pars(dist, logfrailtypar, pvfm)
}
\arguments{
\item{dist}{One of gamma, stable, pvf}

\item{logfrailtypar}{log theta and log llambda}

\item{pvfm}{The pvfm}
}
\value{
A list with 4 elements: aalpha, ggamma (the parameters of the Laplace transform), llambda (distance parameter) and dist_id.
}
\description{
dist_to_pars
}
\keyword{internal}
