% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfrail.R
\name{dynfrail}
\alias{dynfrail}
\title{Fitting dynamic frailty models with the EM algorithm}
\usage{
dynfrail(formula, data, distribution = dynfrail_distribution(),
  control = dynfrail_control(), model = FALSE, model.matrix = FALSE, ...)
}
\arguments{
\item{formula}{A formula that contains on the left hand side an object of the type \code{Surv}
and on the right hand side a \code{+cluster(id)} statement. Optionally, also a \code{+terminal()} statement
may be added, and then a score test for association between the event process and the result in the specified
column is performed. See details.}

\item{data}{A data frame in which the formula argument can be evaluated}

\item{distribution}{An object as created by \code{\link{dynfrail_dist}}}

\item{control}{An object as created by \code{\link{dynfrail_control}}}

\item{model}{Logical. Should the model frame be returned?}

\item{model.matrix}{Logical. Should the model matrix be returned?}

\item{...}{Other arguments, currently used to warn about deprecated argument names}
}
\description{
Fitting dynamic frailty models with the EM algorithm
}
