% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfrail.R
\name{dynfrail}
\alias{dynfrail}
\title{Fitting dynamic frailty models with the EM algorithm}
\usage{
dynfrail(formula, data, distribution = dynfrail_dist(),
  control = dynfrail_control(), ...)
}
\arguments{
\item{formula}{A formula that contains on the left hand side an object of the type \code{Surv}
and on the right hand side a \code{+cluster(id)} statement. Optionally, also a \code{+terminal()} statement
may be added, and then a score test for association between the event process and the result in the specified
column is performed. See details.}

\item{data}{A data frame in which the formula argument can be evaluated}

\item{distribution}{An object as created by \code{\link{dynfrail_dist}}}

\item{control}{An object as created by \code{\link{dynfrail_control}}}

\item{...}{Other arguments, currently used to warn about deprecated argument names}
}
\value{
A \code{dynfrail} object that contains the following fields:
\item{coefficients}{A named vector of the estimated regression coefficients}
\item{hazard}{The breslow estimate of the baseline hazard at each event time point, in chronological order}
\item{imat}{Fisher's information matrix corresponding to the coefficients and hazard, assuming \eqn{\theta, \lambda} constant}
\item{logtheta}{The point estimate of the logarithm of the frailty parameter \eqn{\theta}. See details.}
\item{loglambda}{The point estimate of the logarithm of the autocorrelation parameter \eqn{\lambda}. See details.}
\item{frail}{A \code{data.frame} containing the variables: \code{id} (cluster id), \code{interval} (for piecewise constant frailty, the label of the interval
on which the frailty is constant), \code{Y} (a \code{Surv} object which determines a starting and a stopping time for each row), \code{frail} (the empirical
Bayes estimates of the piecewise constant frailty corresponding to that specific cluster and that specific time period)}
\item{tev}{The time points of the events in the data set, this is the same length as hazard}
\item{loglik}{A vector of length two with the log-likelihood of the starting Cox model
and the maximized log-likelihood}
\item{formula}{The original formula argument}
\item{distribution}{The original distribution argument}
\item{control}{The original control argument}
}
\description{
Fitting dynamic frailty models with the EM algorithm
}
